Copy and Resend buttons are relative to group
